mcp-client/pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "langchain>=1.2.15",
    "langchain-mcp-adapters>=0.2.2",
    "langchain-mistralai>=1.1.2",
    "langchain-openai>=1.2.1",
    "python-dotenv>=1.2.2",
    "streamlit>=1.56.0",
    "tornado>=6.5.5",
]
